elf32: add 32-bit ELF "hello world" example
One issue remains. The .exit entry of the .rela.plt section
seem to be misaligned, or the info field is incorrectly encoded.

Running the executable correctly prints "hello world" to standard
output, but when linking exit, the linker stops and prints an
error message reporting an inconsistency in the info field of the
GOT relocation entry of exit.

	u@x1 ~/D/d/elf32> ./elf32
	hello world
	Inconsistency detected by ld.so: dl-runtime.c: 80: _dl_fixup: Assertion `ELFW(R_TYPE)(reloc->r_info) == ELF_MACHINE_JMP_SLOT' failed!

If anyone knowns what may be at fault here, please reach out to
me. I've been at this for some time, and can't quite figure it out.