elf32: fix bug in dynamic symbol resolution
We used the incorrect argument to dl_resolve.
Prior to this commit we incorrectly used the
index in the dynamic symbol table, whereas
we should have used the offset of in the
relocation table of each symbol.

Now, "hello world" is correctly printed and
the command exists with status code 42.